ath10k: Enable power-save hack special control.
This lets one set up an AP to ignore power-save requests
and/or force all frames to have the 'more-data' flag set.

This may be useful in certain cases where users want to
send lots of near-real-time data over multicast..but it
also may cause instability since it is effectively breaking
the power-save specifications.  This feature is disabled
by default, and requires special CT firmware to support
the feature.

Signed-off-by: Ben Greear <greearb@candelatech.com>