wifi: iwlwifi: mld: Allow scanning non PSC channels.
I put a 6ghz only AP, 40Mhz wide, on 6Ghz channel 1 (non-psc).
It is not doing any colocated stuff or beaconning on 2.4/5.

be200 would not scan it.

With this patch, it will scan and associate.  I also patched my
supplicant to disable NL80211_SCAN_FLAG_COLOCATED_6GHZ flag on
every 5th scan.  Only with this flag disabled will be200 now
scan on non-psc channels.

My approach is to relax the only-scan-psc-channels logic in iwlwifi
when NL80211_SCAN_FLAG_COLOCATED_6GHZ is not set.

The reason I think this might be valid approach (the part about when
it is not set...)

@NL80211_SCAN_FLAG_COLOCATED_6GHZ: scan for collocated APs reported by
 *      2.4/5 GHz APs. When the flag is set, the scan logic will use the
 *      information from the RNR element found in beacons/probe responses
 *      received on the 2.4/5 GHz channels to actively scan only the 6GHz
 *      channels on which APs are expected to be found. Note that when not set,
 *      the scan logic would scan all 6GHz channels, but since transmission of
 *      probe requests on non-PSC channels is limited, it is highly likely that
 *      these channels would passively be scanned. Also note that when the flag
 *      is set, in addition to the colocated APs, PSC channels would also be
 *      scanned if the user space has asked for it.

Signed-off-by: Ben Greear <greearb@candelatech.com>